Revisando código e implentando correções

+++ src/controller/usuarioController.js
@@ -44,13 +44,13 @@ const usuarioController = {
 	},
 
 	update: async (request, response) => {
-		
+
 		const { cargo, nome_documento, senha, cpf, email, nome_social, data_nascimento } = request.body;
 		const passwordCriptografado = bcrypt.hashSync(senha, 10);
-		
+
 		const { idUsuario } = request.params;
 		const { id, role } = request.usuario;
-		
+
 		const listaErros = validationResult(request);
 		if (!listaErros.isEmpty()) {
 			return response.json(listaErros.errors);
@@ -61,7 +61,7 @@ const usuarioController = {
 				email
 			}
 		});
-		
+
 		const usuario = await Usuario.findByPk(idUsuario);
 
 		if (emailResult) {
@@ -76,11 +76,11 @@ const usuarioController = {
 				email,
 				nome_social,
 				data_nascimento,
-				roles:cargo
+				roles: cargo
 			},
-			{
-				where: { id: idUsuario }
-			}
+				{
+					where: { id: idUsuario }
+				}
 			);
 			return response.json('Dados atualizados');
 		}
@@ -89,36 +89,41 @@ const usuarioController = {
 
 	delete: async (request, response) => {
 
-		const { idUsuario } = request.params;
+		const { idUsuario, deleteUsuario } = request.params;
 		const { id, role } = request.usuario;
 
 		const usuario = await Usuario.findByPk(idUsuario);
 
-		if (usuario.id == id || role == 'Administrador') {
+		if (usuario.id == id && role == 'Administrador') {
+
+			const verifyUsuario = await Usuario.findByPk(deleteUsuario)
+
+			if (!verifyUsuario) { return response.json('Usuário não encontrado') }
+
 			const cursoUsuario = await AlunoCurso.findOne({
 				where: {
-					usuario_id: idUsuario
+					usuario_id: deleteUsuario
 				}
 			});
-
 
 			if (cursoUsuario) {
 				await AlunoCurso.destroy({
 					where: {
-						usuario_id: idUsuario
+						usuario_id: deleteUsuario
 					}
 				});
 			}
-            
+
 			if (usuario) {
 				await Usuario.destroy({
 					where: {
-						id: usuario
+						id: deleteUsuario
 					}
 				});
 			}
 			return response.json('Dados deletados');
 		}
+		return response.json('Você não tem autorização para esta função');
 	}
 };
 module.exports = usuarioController;
